requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
debugpy==1.8.2
decorator==5.1.1
Django==5.0.6
django-celery-results==2.5.1
django-redis==5.4.0
djangorestframework==3.15.2
docker==7.1.0
exceptiongroup==1.2.1
idna==3.7
iniconfig==2.0.0
kombu==5.3.7
packaging==24.1
pluggy==1.5.0
prompt_toolkit==3.0.47
psutil==6.0.0
psycopg2-binary==2.9.9
py==1.11.0
pytest==8.2.2
pytest-celery==1.0.0
pytest-django==4.8.0
pytest-docker-tools==3.1.3
python-dateutil==2.9.0.post0
redis==5.0.7
requests==2.32.3
retry==0.9.2
six==1.16.0
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
